COMP: Large 64 bit integers loose precision in double
warning: implicit conversion from 'unsigned long' to 'double'
changes value from 18446744073709551615 to 18446744073709551616 [-Wimplicit-int-float-conversion]

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=28811
git-svn-id: http://svn.slicer.org/Slicer4/trunk@28811 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/vtkITK/vtkITKArchetypeImageSeriesReader.cxx
@@ -860,18 +860,18 @@ int vtkITKArchetypeImageSeriesReader::RequestInformation(
           }
         if ( imageIO->GetComponentType() == itk::ImageIOBase::INT )
           {
-          min = std::numeric_limits<int32_t>::min() < min ? std::numeric_limits<int32_t>::min() : min;
-          max = std::numeric_limits<int32_t>::max() > max ? std::numeric_limits<int32_t>::max() : max;
+          min = static_cast<double>(std::numeric_limits<int32_t>::min() < min ? std::numeric_limits<int32_t>::min() : min);
+          max = static_cast<double>(std::numeric_limits<int32_t>::max() > max ? std::numeric_limits<int32_t>::max() : max);
           }
         if ( imageIO->GetComponentType() == itk::ImageIOBase::ULONG )
           { // note that on windows ULONG is only 32 bit
-          min = std::numeric_limits<uint64_t>::min() < min ? std::numeric_limits<uint64_t>::min() : min;
-          max = std::numeric_limits<uint64_t>::max() > max ? std::numeric_limits<uint64_t>::max() : max;
+          min = static_cast<double>(std::numeric_limits<uint64_t>::min() < min ? std::numeric_limits<uint64_t>::min() : min);
+          max = static_cast<double>(std::numeric_limits<uint64_t>::max() > max ? std::numeric_limits<uint64_t>::max() : max);
           }
         if ( imageIO->GetComponentType() == itk::ImageIOBase::LONG )
           { // note that on windows LONG is only 32 bit
-          min = std::numeric_limits<int64_t>::min() < min ? std::numeric_limits<int64_t>::min() : min;
-          max = std::numeric_limits<int64_t>::max() > max ? std::numeric_limits<int64_t>::max() : max;
+          min = static_cast<double>(std::numeric_limits<int64_t>::min() < min ? std::numeric_limits<int64_t>::min() : min);
+          max = static_cast<double>(std::numeric_limits<int64_t>::max() > max ? std::numeric_limits<int64_t>::max() : max);
           }
         if ( imageIO->GetComponentType() == itk::ImageIOBase::FLOAT )
           {
@@ -901,7 +901,7 @@ int vtkITKArchetypeImageSeriesReader::RequestInformation(
           {
           scalarType = VTK_UNSIGNED_INT;
           }
-        else if( max <= std::numeric_limits<uint64_t>::max() )
+        else if( max <= static_cast<double>(std::numeric_limits<uint64_t>::max()) )
           {
           scalarType = VTK_UNSIGNED_LONG;
           }
